app/requirements.txt
streamlit
attackcti
mitreattack-python
pwntools
pandas
stix2
ollama
python-dotenv
langchain
langchain-community
mcp
tavily-python
httpx
